Multiple updates

+++ src/lib/vpc.py
@@ -14,7 +14,6 @@ class VPC(object):
     classdocs
     '''
 
-
     def __init__(self, name, cidr_block):
         '''
         Constructor
@@ -26,11 +25,29 @@ class VPC(object):
         self.security_groups = []
         self.routing_tables = []
         self.cidr_block = cidr_block
+        self.key = None
         self.subnet = None
         
         try:
             self.vpc_conn = vpc.VPCConnection()
             self.vpc = self.vpc_conn.create_vpc(self.cidr_block)
+            
+            # Create a subnet in a vpc
+            self.subnet = self.vpc_conn.create_subnet(self.vpc.id,
+                                                      self.cidr_block)
+            
+            # Create and attach an internet gateway
+            self.gateway = self.vpc_conn.create_internet_gateway()
+            self.vpc_conn.attach_internet_gateway(self.gateway.id, self.vpc.id)
+            
+            all_route_tables = self.vpc_conn.get_all_route_tables(None,
+                                               filters=[("vpc-id", self.vpc.id)])
+            
+            # pick any first route table and add default route for the internet
+            if all_route_tables is not None:
+                self.vpc_conn.create_route(all_route_tables[0].id, "0.0.0.0/0",
+                                           self.gateway.id)
+            
         except:
             print "Exception occurred"
             raise
@@ -41,8 +58,6 @@ class VPC(object):
                             private_ip_address=None,
                             subnet_id=None, instance_profile_name=None):
         try:
-            self.subnet = self.vpc_conn.create_subnet(self.vpc.id, self.cidr_block)
-            print "subnet id is: " , self.subnet.id
             vm = aws_vm.AWSVm(vm_name, ami_name, instance_type, region,
                               key_name, placement, security_groups,
                               user_data, private_ip_address, self.subnet.id,
@@ -70,7 +85,11 @@ class VPC(object):
             print self.vpc.id
             sec_group = security_group.SecurityGroup(name, description, region,
                                                      self.vpc.id)
-            self.security_groups.append(sec_group)
+            for group in self.security_groups:
+                if group.name == sec_group.name:
+                    break
+            else:
+                self.security_groups.append(sec_group)
             #time.sleep(10)
             for rule in rules:
                 sec_group.add_inbound_rule(rule)
@@ -88,6 +107,17 @@ class VPC(object):
         except:
             print "Exception"
             raise
+       
+        
+    def add_rule_in_security_group(self, security_group_name, rule):
+        
+        try:
+            for sec_group in self.security_groups:
+                if sec_group.name == security_group_name:
+                    sec_group.add_inbound_rule(rule)
+        except:
+            print "VPC: add_rule_in_security_group"
+            raise
         
     def get_security_group_id(self, name):
         for group in self.security_groups:
@@ -103,14 +133,20 @@ class VPC(object):
     # xxx disassociate elastic ip subnets etc
     def __del__(self):
         try:
-            for group in self.security_groups:
-                group.delete_security_group()
-                
             for vm in self.aws_instances:
                 vm.terminate()
                 
+            time.sleep(60)
+                
             if self.subnet is not None: 
                 self.vpc_conn.delete_subnet(self.subnet.id)
+                
+            self.vpc_conn.detach_internet_gateway(self.gateway.id, self.vpc.id)
+            self.vpc_conn.delete_internet_gateway(self.gateway.id)
+            
+            for group in self.security_groups:
+                if group.name != "default":
+                    group.delete_security_group()
             #self.security_groups = []
             #self.aws_instances = []
             self.vpc_conn.delete_vpc(self.vpc.id)
